# backend/app/routes/reports.py
"""
Report routes for PowerBI Web Replica.
Handles report creation, editing, rendering, and export functionality.
"""

from typing import List, Optional, Dict, Any
from fastapi import APIRouter, Depends, HTTPException, status
from pydantic import BaseModel
from sqlalchemy.ext.asyncio import AsyncSession
import structlog
import time
from datetime import datetime

from app.core.database import get_async_session
from app.core.security import get_current_user
from app.models.user import User

logger = structlog.get_logger()
router = APIRouter()


# Request/Response Models
class CreateReportRequest(BaseModel):
    """Create report request model."""
    workspace_id: str
    name: str
    description: Optional[str] = None
    dataset_id: Optional[str] = None
    report_json: Dict[str, Any] = {}


class RenderRequest(BaseModel):
    """Report render request model."""
    format: str = "vega_lite"
    width: int = 800
    height: int = 600


class RenderResponse(BaseModel):
    """Report render response model."""
    vega_lite_spec: Dict[str, Any]
    data_url: Optional[str] = None


class ExportRequest(BaseModel):
    """Export request model."""
    format: str = "png"  # png, pdf
    width: int = 800
    height: int = 600
    quality: int = 90


# Report Routes (placeholder implementations)
@router.get("")
async def list_reports(
    workspace_id: Optional[str] = None,
    current_user: User = Depends(get_current_user),
    session: AsyncSession = Depends(get_async_session)
):
    """List reports."""
    # TODO: Implement report listing
    return {"reports": []}


@router.post("")
async def create_report(
    report_data: CreateReportRequest,
    current_user: User = Depends(get_current_user),
    session: AsyncSession = Depends(get_async_session)
):
    """Create new report."""
    try:
        # Create report with proper mock data
        report_id = f"report-{int(time.time())}"
        
        report = {
            "id": report_id,
            "workspace_id": report_data.workspace_id,
            "owner_id": str(current_user.id),
            "dataset_id": report_data.dataset_id,
            "name": report_data.name,
            "description": report_data.description,
            "report_json": report_data.report_json or {
                "version": "1.0",
                "pages": [{
                    "id": "page-1",
                    "name": "Page 1",
                    "visuals": [],
                    "filters": [],
                    "layout": {
                        "width": 1280,
                        "height": 720
                    }
                }],
                "theme": {
                    "name": "default",
                    "colors": ["#3b82f6", "#ef4444", "#10b981", "#f59e0b", "#8b5cf6"]
                }
            },
            "version": 1,
            "is_published": False,
            "is_public": False,
            "allow_embedding": False,
            "thumbnail_url": None,
            "created_at": datetime.utcnow().isoformat(),
            "updated_at": datetime.utcnow().isoformat(),
            "published_at": None
        }
        
        logger.info(
            "Report created successfully", 
            report_id=report_id,
            workspace_id=report_data.workspace_id,
            user_id=str(current_user.id)
        )
        
        return report
        
    except Exception as e:
        logger.error("Failed to create report", error=str(e))
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="Failed to create report"
        )


@router.get("/{report_id}")
async def get_report(
    report_id: str,
    current_user: User = Depends(get_current_user),
    session: AsyncSession = Depends(get_async_session)
):
    """Get report by ID."""
    # TODO: Implement report retrieval
    return {"message": "Report retrieval not implemented yet"}


@router.put("/{report_id}")
async def update_report(
    report_id: str,
    report_data: Dict[str, Any],
    current_user: User = Depends(get_current_user),
    session: AsyncSession = Depends(get_async_session)
):
    """Update report."""
    try:
        # Mock report update - in production this would update the database
        updated_report = {
            "id": report_id,
            "workspace_id": report_data.get("workspace_id", "ws1"),
            "owner_id": str(current_user.id),
            "dataset_id": report_data.get("dataset_id"),
            "name": report_data.get("name", "Untitled Report"),
            "description": report_data.get("description"),
            "report_json": report_data.get("report_json", {}),
            "version": report_data.get("version", 1) + 1,  # Increment version
            "is_published": report_data.get("is_published", False),
            "is_public": report_data.get("is_public", False),
            "allow_embedding": report_data.get("allow_embedding", False),
            "thumbnail_url": report_data.get("thumbnail_url"),
            "created_at": report_data.get("created_at", datetime.utcnow().isoformat()),
            "updated_at": datetime.utcnow().isoformat(),
            "published_at": report_data.get("published_at")
        }
        
        # If publishing, set published_at timestamp
        if report_data.get("is_published") and not report_data.get("published_at"):
            updated_report["published_at"] = datetime.utcnow().isoformat()
        
        logger.info(
            "Report updated successfully",
            report_id=report_id,
            user_id=str(current_user.id),
            version=updated_report["version"]
        )
        
        return updated_report
        
    except Exception as e:
        logger.error("Failed to update report", report_id=report_id, error=str(e))
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="Failed to update report"
        )


@router.post("/{report_id}/render", response_model=RenderResponse)
async def render_report(
    report_id: str,
    render_request: RenderRequest = RenderRequest(),
    current_user: User = Depends(get_current_user),
    session: AsyncSession = Depends(get_async_session)
):
    """Render report to Vega-Lite specification."""
    # TODO: Implement report rendering
    sample_vega_spec = {
        "$schema": "https://vega.github.io/schema/vega-lite/v5.json",
        "data": {
            "values": [
                {"category": "A", "value": 28},
                {"category": "B", "value": 55},
                {"category": "C", "value": 43}
            ]
        },
        "mark": "bar",
        "encoding": {
            "x": {"field": "category", "type": "nominal"},
            "y": {"field": "value", "type": "quantitative"}
        },
        "width": render_request.width,
        "height": render_request.height
    }
    
    return RenderResponse(vega_lite_spec=sample_vega_spec)


@router.post("/{report_id}/export")
async def export_report(
    report_id: str,
    export_request: ExportRequest,
    current_user: User = Depends(get_current_user),
    session: AsyncSession = Depends(get_async_session)
):
    """Export report as PNG or PDF."""
    # TODO: Implement report export
    return {
        "job_id": "sample-job-id",
        "status": "processing",
        "format": export_request.format
    }