# openapi.yml
openapi: 3.0.3
info:
  title: PowerBI Web Replica API
  description: |
    A comprehensive API for a web-based data visualization platform that recreates 
    Power BI Desktop functionality with select Tableau features.
    
    ## Features
    - Interactive report and dashboard design
    - Multiple data source connectors (CSV, PostgreSQL, etc.)
    - Drag-and-drop visualization builder
    - Real-time chart rendering with Vega-Lite
    - Offline mode support with IndexedDB
    - Export capabilities (PNG, PDF)
    - Role-based access control
    
  version: 1.0.0
  contact:
    name: PowerBI Web Replica
    url: https://github.com/powerbi-web-replica
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.powerbi-replica.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /api/reports:
    get:
      summary: List reports
      tags: [Reports]
      parameters:
        - name: workspace_id
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of reports
          content:
            application/json:
              schema:
                type: object
                properties:
                  reports:
                    type: array
                    items:
                      $ref: '#/components/schemas/Report'
                  total:
                    type: integer

    post:
      summary: Create new report
      tags: [Reports]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workspace_id, name, report_json]
              properties:
                workspace_id:
                  type: string
                  format: uuid
                name:
                  type: string
                  example: Sales Dashboard
                description:
                  type: string
                report_json:
                  type: object
                  description: Report definition in internal JSON format
      responses:
        '201':
          description: Report created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
  /api/reports/{report_id}:
    get:
      summary: Get report
      tags: [Reports]
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Report details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'

    put:
      summary: Update report
      tags: [Reports]
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                report_json:
                  type: object
      responses:
        '200':
          description: Report updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
  /api/reports/{report_id}/render:
    post:
      summary: Render report to Vega-Lite specification
      tags: [Reports]
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Vega-Lite specification for rendering
          content:
            application/json:
              schema:
                type: object
                properties:
                  vega_lite_spec:
                    type: object
                    description: Complete Vega-Lite specification
                  data_url:
                    type: string
                    description: URL to fetch data for the chart
  /api/reports/{report_id}/export:
    post:
      summary: Export report as PNG or PDF
      tags: [Reports]
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [format]
              properties:
                format:
                  type: string
                  enum: [png, pdf]
                  example: png
                width:
                  type: integer
                  default: 800
                  example: 1200
                height:
                  type: integer
                  default: 600
                  example: 800
                quality:
                  type: integer
                  minimum: 1
                  maximum: 100
                  default: 90
      responses:
        '202':
          description: Export job started
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                    format: uuid
                  status:
                    type: string
                    example: processing

  # Dashboard Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Report:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        report_json:
          type: object
          description: Internal report definition
        owner_id:
          type: string
          format: uuid
        version:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...